# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LinguaAI Language Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  title: string;
  language: string;
  level: string;
  skill: string;
  content: string;
  /** @nullable */
  audioUrl?: string | null;
  duration: number;
  xp: number;
  isCompleted?: boolean;
  description?: string;
  objectives?: string[];
  tags?: string[];
}