# lib/api-zod/src/generated/types/levelBreakdown.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LinguaAI Language Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LevelBreakdown {
  level: string;
  language: string;
  readingProgress: number;
  listeningProgress: number;
  writingProgress: number;
  totalLessons: number;
  completedLessons: number;
}