# lib/api-zod/src/generated/types/progressInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LinguaAI Language Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressInput {
  language: string;
  lessonId: number;
  xpEarned: number;
  skill: string;
  score?: number;
}