# lib/api-zod/src/generated/types/textCorrection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LinguaAI Language Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TextCorrection {
  original: string;
  corrected: string;
  explanation: string;
}