# lib/api-zod/src/generated/types/userProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LinguaAI Language Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProgress {
  language: string;
  level: string;
  totalXp: number;
  lessonsCompleted: number;
  currentStreak: number;
  wordsLearned: number;
  readingScore?: number;
  listeningScore?: number;
  writingScore?: number;
  lastActivity?: string;
}